added scalar and one SSE implementation of set_difference
The SSE implementation is nearly the same as for set intersection. It's
checking for common elements in both lists. Instead of writing them out
immediately like for intersection, the non-common elements are written out when
the first list is stepped forward to the next 4 elements. Only then we can be
sure that the current vector elements of list1 have no more matches in list2.