fixed deduplication across vector boundaries in set union
The AVX512 set union implementation uses a bitonic merge network to
sort two vectors. At the end the first vector contains the lower 16
32bit-elements and the second the higher ones. The first vector is
stored without the duplicates and the second vector is used for the
next iteration.

Duplicate elements can be easily found with vpconflictd inside one
vector. There can also be a duplicate across the vector bounderies, one
entry in both vectors (at the end of the first vector and at the start
of the second). This corner case is now handled by keeping track of the
last element of the previously saved vector. As there's no scalar
extract instruction, compressstore is used to extract the first and
last element. The last element is stored for the next iteration.