fixed AVX512 set_difference with vpconflictd
Using unsigned for the mask handling in the tail instead of uint16_t
to get around a weird problem in gcc. With uint16_t gcc (6.3) favors
to do shift then in the mask registers which leads to silly code:
move from register to mask register, shift the mask register and move
back to general purpose register. With unsigned it's using the normal
shift again.